+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "account-abstraction",
-  "version": "0.1.0",
+  "version": "0.1.0-a1",
   "description": "",
-  "main": "dist/index.js",
+  "main": "dist/src/index.js",
   "scripts": {
     "clean": "rm -rf cache artifacts typechain",
     "compile": "./scripts/hh-wrapper compile && yarn run makeUserOps",
@@ -12,35 +12,44 @@
     "test-dev": "hardhat test --network dev",
     "ci": "hardhat compile && hardhat test",
     "runop": "hardhat run src/runop.ts ",
-    "runop3": "hardhat run src/runop3.ts "
+    "runop3": "hardhat run src/runop3.ts ",
+    "prepack": "yarn clean && yarn compile && tsc && yarn missing-d.ts-files",
+    "missing-d.ts-files": "cp typechain/*.d.ts dist/typechain/; touch dist/typechain/EntryPoint.js dist/typechain/SimpleWallet.js",
+    "tsc": "tsc"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "@nomiclabs/hardhat-ethers": "^2.0.2",
+    "@nomiclabs/hardhat-etherscan": "^2.1.6",
     "@nomiclabs/hardhat-waffle": "^2.0.1",
+    "@openzeppelin/contracts": "^4.2.0",
     "@typechain/ethers-v5": "^7.0.1",
+    "@typechain/hardhat": "^2.3.0",
     "@types/chai": "^4.2.21",
+    "@types/mocha": "^9.0.0",
     "@types/node": "^16.4.12",
     "chai": "^4.3.4",
     "ethereum-waffle": "^3.4.0",
-    "ethers": "^5.4.2",
     "hardhat": "^2.6.6",
-    "ts-generator": "^0.1.1",
-    "ts-node": "^10.1.0",
-    "typechain": "^5.1.2"
-  },
-  "dependencies": {
-    "@nomiclabs/hardhat-etherscan": "^2.1.6",
-    "@openzeppelin/contracts": "^4.2.0",
-    "@typechain/hardhat": "^2.3.0",
-    "@types/mocha": "^9.0.0",
-    "ethereumjs-util": "^7.1.0",
-    "ethereumjs-wallet": "^1.0.1",
     "hardhat-deploy": "^0.9.3",
     "hardhat-deploy-ethers": "^0.3.0-beta.11",
     "source-map-support": "^0.5.19",
+    "ts-generator": "^0.1.1",
+    "ts-node": "^10.1.0",
+    "typechain": "^5.1.2",
     "typescript": "^4.3.5"
-  }
+  },
+  "dependencies": {
+    "debug": "^4.3.2",
+    "ethereumjs-util": "^7.1.0",
+    "ethers": "^5.5.1"
+  },
+  "files": [
+    "dist",
+    "src",
+    "typechain",
+    "README.md"
+  ]
 }
