+++ scripts/makeUserOps
@@ -4,4 +4,4 @@ perl -ne 'next unless /struct UserOperation/../}/; \
         s/\s*struct/import * as typ from ".\/solidityTypes";\n\nexport interface/; \
         s/\s*(\w+)\s+(\w+);/  $2: typ.$1/; \
         s/\s*}/}/; \
-        print' contracts/UserOperation.sol > test/UserOperation.ts
+        print' contracts/UserOperation.sol > src/userop/UserOperation.ts
