+++ tsconfig.json
@@ -17,7 +17,6 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true,
     "outDir": "dist",
     "declaration": true,
     "declarationMap": true,
